refactoring map filter form, part 1
