+++ src/Infrastructure/Controller/Map/Fragments/GetMapDataController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure\Controller\Map;
+namespace App\Infrastructure\Controller\Map\Fragments;
 
 use App\Domain\Regulation\Repository\LocationRepositoryInterface;
 use App\Infrastructure\Controller\DTO\MapFilterDTO;
@@ -12,8 +12,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\UX\Turbo\TurboBundle;
 
-final class GetMapFilterController
+final class GetMapDataController
 {
     public function __construct(
         private \Twig\Environment $twig,
@@ -24,8 +25,8 @@ final class GetMapFilterController
     }
 
     #[Route(
-        '/_get_map_filter',
-        name: 'get_map_filter',
+        '/map/fragments/data',
+        name: 'fragment_map_data',
         methods: ['GET'],
     )]
     public function __invoke(Request $request): Response
@@ -34,32 +35,24 @@ final class GetMapFilterController
         $form = $this->formFactory->create(
             type: MapFilterFormType::class,
             data: $dto,
-            options: [
-                'action' => $this->router->generate('get_map_filter'),
-                'method' => 'GET',
-                'attr' => [
-                    'data-turbo-action' => 'replace',
-                ],
-            ],
         );
-
-        $form->handleRequest($request); // auto-fill the form with the query parameters from the URL
-
-        $locationsAsGeoJson = $this->locationRepository->findAllForMapAsGeoJSON(
+        //$form->handleRequest($request); // auto-fill the form with the query parameters from the URL
+        $locationsAsGeoJsonAsText = $this->locationRepository->findAllForMapAsGeoJSON(
             $dto->category === 'permanents_only',
             $dto->category === 'temporaries_only',
             $dto->displayFutureRegulations === '1',
             $dto->displayPastRegulations === '1',
         );
 
+        $request->setRequestFormat(TurboBundle::STREAM_FORMAT); // for the response (!)
         return new Response(
             $this->twig->render(
-                name: 'map/map_filter.html.twig',
+                name: 'map/fragments/map_data.stream.html.twig',
                 context: [
-                    'locationsAsGeoJson' => $locationsAsGeoJson,
-                    'form' => $form->createView(),
+                    'locationsAsGeoJson' => $locationsAsGeoJsonAsText,
                 ],
             ),
         );
+
     }
 }
