+++ templates/map/fragments/map_data.stream.html.twig
@@ -1,0 +1,5 @@
+<turbo-stream action="update" target="locations_as_geojson">
+  <template>
+    {{ locationsAsGeoJson }}
+  </template>
+</turbo-stream>
